Remove ability to change bind address & port for HTTP server
While this is handy, trying to keep the number of hours invested in this
down means that I won't make the Vue.js app be able to automatically
account for changes in the port. Rather than having & documenting a
feature that _will_ break the Vue.js app, I'm removing the ability to
change the bind address & port from an environment variable.